+++ Ya/Algebra.hs
@@ -1,3 +1,4 @@
 module Ya.Algebra (module Exports) where
 
 import Ya.Algebra.Abstract as Exports
+import Ya.Algebra.Definition as Exports