+++ CHANGELOG.md
@@ -18,3 +18,7 @@
 * Define `Recursive` data type
 * Define `U_UU_UUU_UUUU_T_TT_I_II_III` data type
 * Define `R_U_I_T_I` data type
+* Define `Transformation` type class
+* Define `unwrap` expression
+* Define `wrap` expression
+* Define `transform` expression